+++ astropy/io/fits/fitsrec.py
@@ -846,7 +846,15 @@ class FITS_rec(np.recarray):
         dummy = np.char.ljust(field, format.width)
         dummy = np.char.replace(dummy, encode_ascii('D'), encode_ascii('E'))
         null_fill = encode_ascii(str(ASCIITNULL).rjust(format.width))
+
+        # Convert all fields equal to the TNULL value (nullval) to empty fields.
+        # TODO: These fields really should be conerted to NaN or something else undefined.
+        # Currently they are converted to empty fields, which are then set to zero.
         dummy = np.where(np.char.strip(dummy) == nullval, null_fill, dummy)
+
+        # always replace empty fields, see https://github.com/astropy/astropy/pull/5394
+        if nullval != b'':
+            dummy = np.where(np.char.strip(dummy) == b'', null_fill, dummy)
 
         try:
             dummy = np.array(dummy, dtype=recformat)